+++ persistence/src/main/java/com/redberrystudios/whatsfordinner/checklist/ChecklistMongoRepository.java
@@ -1,21 +1,23 @@
 package com.redberrystudios.whatsfordinner.checklist;
 
-import static com.mongodb.client.model.Filters.eq;
-import static com.mongodb.client.model.Filters.in;
-
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.UpdateOptions;
 import com.redberrystudios.whatsfordinner.MongoRepository;
 import com.redberrystudios.whatsfordinner.group.GroupEntity;
 import com.redberrystudios.whatsfordinner.group.GroupMongoRepository;
-import java.util.ArrayList;
-import java.util.List;
+import org.bson.BsonValue;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.mongodb.client.model.Filters.eq;
+import static com.mongodb.client.model.Filters.in;
+
 @Repository
-public class ChecklistMongoRepository extends MongoRepository {
+public class ChecklistMongoRepository extends MongoRepository<ChecklistEntity, Long> {
 
   private static final String COLLECTION_NAME = "checklists";
 
@@ -31,21 +33,45 @@ public class ChecklistMongoRepository extends MongoRepository {
     this.groupMongoRepository = groupMongoRepository;
   }
 
-  public void save(ChecklistEntity checklist) {
-    collection.replaceOne(eq("_id", checklist.getId()),
+  public Long save(ChecklistEntity checklist) {
+    if (checklist == null) {
+      throw new IllegalArgumentException("ChecklistEntity to save is null!");
+    }
+
+    BsonValue upsertId = collection.replaceOne(eq("_id", checklist.getId()),
         checklist,
-        new UpdateOptions().upsert(true));
+        new UpdateOptions().upsert(true))
+        .getUpsertedId();
+
+    if (upsertId != null) {
+      return upsertId.asInt64().getValue();
+    } else {
+      return checklist.getId();
+    }
   }
 
-  public void delete(ChecklistEntity checklist) {
+  public Long delete(ChecklistEntity checklist) {
+    if (checklist == null) {
+      throw new IllegalArgumentException("ChecklistEntity to delete is null!");
+    }
+
     collection.deleteOne(eq("_id", checklist.getId()));
+    return checklist.getId();
   }
 
   public ChecklistEntity find(Long checklistId) {
+    if (checklistId == null) {
+      return null;
+    }
+
     return collection.find(eq("_id", checklistId)).first();
   }
 
   public ChecklistEntity findByGroup(Long groupId, Long checkListId) {
+    if (groupId == null || checkListId == null) {
+      return null;
+    }
+
     GroupEntity groupEntity = groupMongoRepository.find(groupId);
 
     if (groupEntity.getChecklists().contains(checkListId)) {
@@ -56,6 +82,10 @@ public class ChecklistMongoRepository extends MongoRepository {
   }
 
   public List<ChecklistEntity> findAllByGroup(Long groupId) {
+    if (groupId == null) {
+      return null;
+    }
+
     GroupEntity groupEntity = groupMongoRepository.find(groupId);
 
     return collection.find(in("_id", groupEntity.getChecklists()))
